requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.3.0
beautifulsoup4==4.12.3
bleach==6.2.0
boto3==1.36.11
certifi==2024.12.14
charset-normalizer==3.4.1
comm==0.2.2
contourpy==1.3.1
crispy-bootstrap4==2024.10
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.5
django-aux==0.0.94
django-bootstrap-datepicker-plus==5.0.5
django-bootstrap4==24.4
django-crispy-forms==2.3
django-extensions==3.2.3
django-filter==24.3
django-markdownify==0.9.5
django-pandas==0.6.7
django-redshift-backend==5.0.0
django-simple-history==3.7.0
django-tables2==2.7.5
djangorestframework==3.15.2
executing==2.1.0
factory_boy==3.3.1
Faker==33.3.1
fastjsonschema==2.21.1
fonttools==4.55.3
idna==3.10
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lightgbm==4.6.0
lxml==5.3.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.1.0
nbclient==0.10.2
nbconvert==7.16.5
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
pip==24.2
platformdirs==4.3.6
plotly==5.24.1
ply==3.11
prompt_toolkit==3.0.48
psutil==6.1.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.10.5
pydantic_core==2.27.2
Pygments==2.19.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
referencing==0.35.1
requests==2.32.3
rpds-py==0.22.3
scikit-learn==1.6.1
six==1.17.0
skforecast==0.14.0
soupsieve==2.6
sqlparse==0.5.3
stack-data==0.6.3
tenacity==9.0.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
xlrd==2.0.1
